First rewrite of API
Some notes/q's: 
To "apply" an action, I copied over code for the .order method int he OrderBook class, but I'm not sure if that’s the right place to look.
Some functions in the original environment.py calculate cost - where should that be happening in this new version?
Looking at the old code, it seems like the orderbooks used dictionaries for the bids and asks or a vector of levels. In this new version, the bid and ask vectors are just lists of Levels, but would dictionaries make more sense?